fixed atoi for working with long int

+++ ft_atoi.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	res;
+	long	res;
 	int	sign;
 	int	sign_exist;
